requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
motor==3.5.0
multidict==6.0.5
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.8.0
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
telebot==0.0.5
typing_extensions==4.10.0
urllib3==2.2.2
uvicorn==0.28.0
yarl==1.9.4
